feat: Polynome.derivee()
la classe Polynome est maintenant définie dans un fichier auxiliaire
